Refactor: Extract static projection math from ConstraintModuleManager
